Add @nordeck/matrix-meetings-calendar package and move dateTimeUtils into it.
