+++ .changeset/cuddly-rings-beam.md
@@ -1,0 +1,6 @@
+---
+'@nordeck/matrix-meetings-widget': patch
+'@nordeck/matrix-meetings-bot': patch
+---
+
+Use `@nordeck/matrix-meetings-calendar` package.
