+++ .changeset/heavy-glasses-shout.md
@@ -1,0 +1,5 @@
+---
+'@nordeck/matrix-meetings-calendar': major
+---
+
+Initial release
