+++ .prettierignore
@@ -2,6 +2,7 @@
 **/coverage/
 matrix-meetings-bot/lib/
 matrix-meetings-bot/storage
+packages/calendar/lib
 
 # Ignore the helm chart
 charts/
