+++ matrix-meetings-bot/src/shared/calendarUtils/getCalendarEnd.ts
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
+import { parseICalDate } from '@nordeck/matrix-meetings-calendar';
 import { max } from 'lodash';
 import { DateTime } from 'luxon';
 import { CalendarEntryDto } from '../../dto/CalendarEntryDto';
-import { parseICalDate } from '../dateTimeUtils';
 import { generateRruleSet } from './generateRruleSet';
 import { isRRuleEntry, isRRuleOverrideEntry, isSingleEntry } from './helpers';
