--- matrix-meetings-bot/src/shared/dateTimeUtils.test.ts
@@ -1,80 +1,0 @@
-/*
- * Copyright 2022 Nordeck IT + Consulting GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { DateTime } from 'luxon';
-import { formatICalDate, parseICalDate, toISOString } from './dateTimeUtils';
-
-describe('parseICalDate', () => {
-  it.each`
-    input                | output
-    ${'20221029T100000'} | ${'2022-10-29T10:00:00.000Z'}
-    ${'20221030T100000'} | ${'2022-10-30T10:00:00.000Z'}
-    ${'20221030T180000'} | ${'2022-10-30T18:00:00.000Z'}
-  `('should parse $input', ({ input, output }) => {
-    expect(parseICalDate({ value: input, tzid: 'UTC' }).toISO()).toBe(output);
-  });
-
-  it.each`
-    input                | output
-    ${'20221029T100000'} | ${'2022-10-29T10:00:00.000+02:00'}
-    ${'20221030T100000'} | ${'2022-10-30T10:00:00.000+01:00'}
-    ${'20221030T180000'} | ${'2022-10-30T18:00:00.000+01:00'}
-  `('should parse $input with Europe/Berlin timezone', ({ input, output }) => {
-    expect(parseICalDate({ value: input, tzid: 'Europe/Berlin' }).toISO()).toBe(
-      output,
-    );
-  });
-});
-
-describe('formatICalDate', () => {
-  it.each`
-    input                     | value
-    ${'2022-10-29T10:00:00Z'} | ${'20221029T100000'}
-    ${'2022-10-30T10:00:00Z'} | ${'20221030T100000'}
-    ${'2022-10-30T18:00:00Z'} | ${'20221030T180000'}
-  `('should format $input', ({ input, value }) => {
-    expect(formatICalDate(new Date(input))).toEqual({
-      value,
-      tzid: 'UTC',
-    });
-  });
-
-  it.each`
-    input                     | value
-    ${'2022-10-29T08:00:00Z'} | ${'20221029T100000'}
-    ${'2022-10-30T09:00:00Z'} | ${'20221030T100000'}
-    ${'2022-10-30T17:00:00Z'} | ${'20221030T180000'}
-  `('should format $input with Europe/Berlin timezone', ({ input, value }) => {
-    expect(formatICalDate(DateTime.fromISO(input), 'Europe/Berlin')).toEqual({
-      value,
-      tzid: 'Europe/Berlin',
-    });
-  });
-});
-
-describe('toISOString', () => {
-  it('should format javascript Date', () => {
-    expect(toISOString(new Date('2022-10-29T08:00:00Z'))).toBe(
-      '2022-10-29T08:00:00Z',
-    );
-  });
-
-  it('should format luxon DateTime', () => {
-    expect(toISOString(DateTime.fromISO('2022-10-29T08:00:00Z'))).toBe(
-      '2022-10-29T08:00:00Z',
-    );
-  });
-});
