+++ matrix-meetings-bot/tsconfig.json
@@ -5,6 +5,7 @@
     "emitDecoratorMetadata": true,
     "module": "commonjs",
     "target": "es6",
+    "lib": ["esnext"],
     "noImplicitAny": true,
     "sourceMap": true,
     "outDir": "./lib",
