+++ matrix-meetings-widget/craco.config.js
@@ -15,7 +15,10 @@
  */
 
 module.exports = {
-  plugins: [{ plugin: require('./scripts/craco/ignoreRRuleWarnings') }],
+  plugins: [
+    { plugin: require('./scripts/craco/ignoreRRuleWarnings') },
+    { plugin: importLocalPackages() },
+  ],
 
   webpack: {
     configure: {
@@ -28,3 +31,31 @@ module.exports = {
     },
   },
 };
+
+/**
+ * Craco plugin for using local packages from the same mono repository.
+ */
+function importLocalPackages() {
+  const path = require('path');
+  const { getLoader, loaderByName } = require('@craco/craco');
+  const absolutePath = path.join(__dirname, '../packages');
+
+  function overrideWebpackConfig({ webpackConfig }) {
+    const { isFound, match } = getLoader(
+      webpackConfig,
+      loaderByName('babel-loader'),
+    );
+    if (isFound) {
+      const include = Array.isArray(match.loader.include)
+        ? match.loader.include
+        : [match.loader.include];
+
+      match.loader.include = include.concat(absolutePath);
+    }
+    return webpackConfig;
+  }
+
+  return {
+    overrideWebpackConfig,
+  };
+}
