+++ matrix-meetings-widget/src/components/meetings/ScheduleMeetingModal/ScheduleMeeting.tsx
@@ -26,6 +26,7 @@ import {
   TextField,
 } from '@mui/material';
 import { unstable_useId as useId, visuallyHidden } from '@mui/utils';
+import { parseICalDate } from '@nordeck/matrix-meetings-calendar';
 import { DateTime } from 'luxon';
 import { ChangeEvent, useCallback, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -33,7 +34,6 @@ import {
   getInitialMeetingTimes,
   isBotUser,
   isRecurringCalendarSourceEntry,
-  parseICalDate,
 } from '../../../lib/utils';
 import {
   Meeting,
