+++ matrix-meetings-widget/src/lib/matrix/model/index.ts
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-export type { CalendarEntry, DateTimeEntry } from './calendarEntry';
 export {
   STATE_EVENT_NORDECK_MEETING_METADATA,
   isValidNordeckMeetingMetadataEvent,
