+++ matrix-meetings-widget/src/lib/utils/calendarUtils/generateRruleSet.ts
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
+import {
+  CalendarEntry,
+  parseICalDate,
+} from '@nordeck/matrix-meetings-calendar';
 import { DateTime } from 'luxon';
 import { RRule, RRuleSet } from 'rrule';
-import { CalendarEntry } from '../../matrix';
-import { parseICalDate } from '../dateTimeUtils';
 import { isFiniteSeries } from './helpers';
 
 export function generateRruleSet(entry: CalendarEntry & { rrule: string }): {
