+++ matrix-meetings-widget/src/lib/utils/index.ts
@@ -15,13 +15,6 @@
  */
 
 export * from './calendarUtils';
-export {
-  formatICalDate,
-  isMeetingSpanningMultipleDays,
-  isMeetingSpanningMultipleYears,
-  parseICalDate,
-  toISOString,
-} from './dateTimeUtils';
 export { generateFilterRange } from './generateFilterRange';
 export type { CalendarViewType } from './generateFilterRange';
 export { getBotUserId, isBotUser } from './getBotUserId';
@@ -30,3 +23,7 @@ export { getWeekdayShift } from './getWeekdayShift';
 export { isDefined } from './isDefined';
 export { isString } from './isString';
 export { localeWeekdays } from './localeWeekdays';
+export {
+  isMeetingSpanningMultipleDays,
+  isMeetingSpanningMultipleYears,
+} from './meetingDateTimeUtils';
