+++ matrix-meetings-widget/src/reducer/meetingsApi/selectors/selectAllMeetingIds.ts
@@ -15,15 +15,11 @@
  */
 
 import { StateEvent } from '@matrix-widget-toolkit/api';
+import { parseICalDate, toISOString } from '@nordeck/matrix-meetings-calendar';
 import { createSelector } from '@reduxjs/toolkit';
 import { DateTime } from 'luxon';
 import { RoomNameEvent, RoomTopicEvent } from '../../../lib/matrix';
-import {
-  calculateCalendarEvents,
-  isString,
-  parseICalDate,
-  toISOString,
-} from '../../../lib/utils';
+import { calculateCalendarEvents, isString } from '../../../lib/utils';
 import { RootState } from '../../../store';
 import {
   isMeetingBreakOutRoom,
