+++ packages/calendar/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@nordeck/matrix-meetings-calendar",
+  "description": "A library sharing calendar utils between meetings widget and bot",
+  "license": "Apache-2.0",
+  "private": true,
+  "version": "0.0.1",
+  "source": "./src/index.ts",
+  "module": "./src/index.ts",
+  "types": "./src/index.ts",
+  "main": "lib/index.js",
+  "scripts": {
+    "docker:build": "echo \"Nothing to build\"",
+    "build": "tsc",
+    "depcheck": "depcheck",
+    "test": "jest --watch",
+    "test:all": "jest --coverage --watchAll=false",
+    "lint": "eslint .",
+    "translate": "echo \"Nothing to translate\"",
+    "generate-disclaimer": "echo \"Nothing to generate\""
+  },
+  "dependencies": {
+    "joi": "^17.11.0",
+    "luxon": "^3.3.0"
+  },
+  "devDependencies": {
+    "@types/luxon": "^3.2.0"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "rootDir": ".",
+    "roots": [
+      "<rootDir>/src"
+    ]
+  }
+}
